Consolidated census_to_gsafac and census_historical_migration apps
