--- backend/census_to_gsafac/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CensusToGsafacConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "census_to_gsafac"
